Modificación rápida
Cambie int por Long para el id que obtiene de la tabla.

+++ PresentacionCISCOE1/src/main/java/ModuloAdministracion/panelComputadorasListado.java
@@ -75,25 +75,25 @@ public class panelComputadorasListado extends javax.swing.JPanel {
     }
 
     private void editar() {
-        int id = this.getIdSeleccionadoTabla();
+        Long id = this.getIdSeleccionadoTabla();
         System.out.println("El id que se va a editar es " + id);
     }
 
     private void eliminar() {
-        int id = this.getIdSeleccionadoTabla();
+        Long id = this.getIdSeleccionadoTabla();
         System.out.println("El id que se va a eliminar es " + id);
     }
 
-    private int getIdSeleccionadoTabla() {
+    private Long getIdSeleccionadoTabla() {
         int indiceFilaSeleccionada = this.tablaComputadoras.getSelectedRow();
         if (indiceFilaSeleccionada != -1) {
             DefaultTableModel modelo = (DefaultTableModel) this.tablaComputadoras.getModel();
             int indiceColumnaId = 0;
-            int idSeleccionado = (int) modelo.getValueAt(indiceFilaSeleccionada,
+            Long idSeleccionado = (Long) modelo.getValueAt(indiceFilaSeleccionada,
                     indiceColumnaId);
             return idSeleccionado;
         } else {
-            return 0;
+            return 0L;
         }
     }
